pyproject.toml
[project]
name = "fastapi-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "imagekitio>=4.2.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
